Revert "Deduplicate words when classifying"
This reverts commit fc6bbda43d8050887b605e8a429a4e67332271cc.

It turns out that this vastly decreases classification performance.
Thinking about this for a few minutes shows why: each word now only
has a tiny fraction of the impact it had before, even if it is
duplicated a lot: Stuffing a message with "Bitcoin, Bitcoin, Bitcoin"
will only be counted as "we've seen 'Bitcoin'" instead of "we've seen
it 3 times".